Handle concurrent downloads of same file better
Previously, in order to download a file `foo/bar.json`, first `foo/bar.json.tmp` would be
downloaded and then it would get renamed to `forr/bar.json`. The problem with this
approach is that it doesn't handle concurrent downloads well.

For example, if two processes are trying to download this file at the same
time, the second one to finish will likely fail with an error of the form:

```
foo/bar.json.tmp: renameFile:renamePath:rename: does not exist (No such file or directory)
```

This is because the faster process already renamed that file.

To fix this, we now download the file to `foo/barZZZZ.json` where `ZZZZ` is guaranteed to
be unique.